fix some code smells

+++ include/server/server.h
@@ -7,7 +7,7 @@
 #define MAX_CON 2
 #define BLOCK_SIZE 4096
 #define HASH_TABLE_SIZE 4096
-struct sockaddr_in *setServerOptions(sa_family_t ip, in_addr_t addr, unsigned int port);
+struct sockaddr_in *setServerOptions(sa_family_t ip, in_addr_t addr, unsigned short port);
 int *startServer(struct sockaddr_in *options);
 void listinLoop(int *server_fd, struct sockaddr_in *options);
 #endif
